+++ thread/costmeasurements/util.h
@@ -1,7 +1,9 @@
 #ifndef __NSTIME_H__
 #define __NSTIME_H__
 
-unsigned long long get_ns_time();
+#include <stdint.h>
+
+uint64_t get_ns_time();
 void errorExit(char *msg);
 void allocateAndUseBuf(int mb);
